Add support for the new RoomList API
Squashed commits:
[8c2625a] Re-enable invites
[339e59b5] Add automatic syncing restarts after entering the terminated state for whatever reason
[b5498e96] Reimplemented room list state management and usage
[8154b0cf] Fix visible rooms range setting after scrolling stops
[5155c44d] Hook up initial loading idicator and session verification banner to room list service states
[700a5a2e] Enable back room subscriptions, timeline listeners and unread notifications / counts
[9df383bc] Enable last messages and timestamps, they don't crash on release builds
[7c4da9da] Add back certain fields now that the "full room" is available
[9f6534a7] Adopt interfaces to new RoomList API, untested as requests don't go to the right URL

+++ ElementX/Sources/Services/Media/MediaUploadingPreprocessor.swift
@@ -190,7 +190,7 @@ struct MediaUploadingPreprocessor {
                                                   mimetype: thumbnailResult.mimeType,
                                                   size: thumbnailSize)
                 
-                let videoInfo = VideoInfo(duration: UInt64(result.duration),
+                let videoInfo = VideoInfo(duration: result.duration,
                                           height: UInt64(result.height),
                                           width: UInt64(result.width),
                                           mimetype: result.mimeType,
@@ -223,7 +223,7 @@ struct MediaUploadingPreprocessor {
             return .failure(.failedProcessingAudio)
         }
         
-        let audioInfo = AudioInfo(duration: UInt64(durationInSeconds * 1000), size: fileSize, mimetype: mimeType)
+        let audioInfo = AudioInfo(duration: durationInSeconds * 1000, size: fileSize, mimetype: mimeType)
         return .success(.audio(audioURL: url, audioInfo: audioInfo))
     }
